Connected backend to MySQL and refactored code to simplify frontend file structure
