--- front-end/src/components/screens/Login.js
@@ -1,12 +1,0 @@
-import React from "react";
-import LoginPage from "../layout/LoginPage";
-
-export default function Login(){
-    return(
-        <>
-            <h1 className="m-5 text-center" >Log In</h1>
-            <LoginPage/>
-        </>
-
-    )
-}
