--- front-end/src/components/screens/Registration.js
@@ -1,12 +1,0 @@
-import React from "react";
-import Forms from "../layout/Forms";
-
-export default function Registration(){
-    return(
-        <>
-            <h1 className="m-5 text-center" >Signup</h1>
-            <Forms/>
-        </>
-        
-    )
-}
